\name{globalOptTests-package}
\alias{globalOptTests-package}
\alias{globalOptTests}
\docType{package}
\title{Objective functions for benchmarking the performance of
  continuous global optimization algorithms}
\description{ This package makes available 50 objective functions for benchmarking the performance of global optimization algorithms based on \code{C} functions described in \cite{M. Montaz Ali, Charoenchai Khompatraporn and Zelda B. Zabinsky, Journal of Global Optimization, 2005, vol. 31, 4, 635-672.}
}
\author{Katharine Mullen <katharine.mullen@stat.ucla.edu>}
\references{
  \code{C} code for the functions implemented here was first published by
  \cite{M. Montaz Ali, Charoenchai Khompatraporn and Zelda B. Zabinsky, Journal
of Global Optimization, 2005, Volume 31, Number 4, Pages 635-672. 
}
The \code{C} implementations were downloaded in a form that was modified from
the paper above, and modified further for incorporation into this
package from a site maintained by GAMS Development Corp. and GAMS
Software GmbH:
\url{http://www.gamsworld.org/performance/selconglobal/selcongloballib.htm}.
}
\keyword{optimization}
