\name{globalOptTests-package}
\alias{globalOptTests-package}
\alias{globalOptTests}
\docType{package}
\title{Objective functions for benchmarking the performance of
  continuous global optimization algorithms}
\description{ This package makes available 50 objective functions for
  benchmarking the performance of global optimization algorithms.
}
\details{
  The \code{C} implementations were downloaded in a form that was modified from
  the paper above, and modified further for incorporation into this
  package from a site maintained by GAMS Development Corp. and GAMS
   Software GmbH:
   \url{http://www.gamsworld.org/performance/selconglobal/selcongloballib.htm}.
}
\author{Katharine Mullen \email{katharine.mullen@stat.ucla.edu} }
\references{
  Montaz Ali, M., Khompatraporn, C. and Zabinsky, Z. B. (2005),
  A Numerical Evaluation of Several Stochastic Algorithms on Selected
  Continuous Global Optimization Test Problems, \emph{Journal
    of Global Optimization}, \bold{31}, 4, 635--672.
  
  Mullen, K. M. (2014), Continuous Global Optimization in R,
  \emph{Journal of Statistical Software}, \bold{60}, 6, 1--45,
  URL \url{http://www.jstatsoft.org/v60/i06/}.

  \url{http://www.gamsworld.org/performance/selconglobal/selcongloballib.htm}
  for the \code{C} source. 
}

\keyword{optimization}
